쓰레기통 등록/카메라 리팩토링
